Language change
I have decided I am going to switch the project over from java to javascript and use sql and html to make it into a website instead of using java swing. I just feel this will make much more sense and make the actual interface of the program much better looking because of the capabilities of using html and css in conjunction.